// src/store/note/note.types.ts
export enum NOTE_ACTION_TYPES {
  SET_CURRENT_NOTES = 'note/SET_CURRENT_NOTES',
  SET_NOTES_COUNT = 'note/SET_NOTES_COUNT',

  ADDING_NOTE_START = 'note/ADDING_NOTE_START',
  ADDING_NOTE_SUCCESS = 'note/ADDING_NOTE_SUCCESS',
  ADDING_NOTE_FAILED = 'note/ADDING_NOTE_FAILED',

  GET_NOTES_START = 'note/GET_NOTES_START',
  GET_NOTES_SUCCESS = 'note/GET_NOTES_SUCCESS',
  GET_NOTES_FAILED = 'note/GET_NOTES_FAILED',

  GET_NOTE_START = 'note/GET_NOTE_START',
  GET_NOTE_SUCCESS = 'note/GET_NOTE_START',
  GET_NOTE_FAILED = 'note/GET_NOTE_FAILED',

  UPDATE_NOTE_START = 'note/UPDATE_NOTE_START',
  UPDATE_NOTE_SUCCESS = 'note/UPDATE_NOTE_SUCCESS',
  UPDATE_NOTE_FAILED = 'note/UPDATE_NOTE_FAILED',

  DELETE_NOTE_START = 'note/DELETE_NOTE_START',
  DELETE_NOTE_SUCCESS = 'note/DELETE_NOTE_SUCCESS',
  DELETE_NOTE_FAILED = 'note/DELETE_NOTE_FAILED',

  CLOSE_NOTE = 'note/CLOSE_NOTE',
}